Update and rename data.ts to expressions.ts
